[real_hwc_wrapper] Do not throw on hwc->set() error
Based on what surfaceflinger does, we should simply ignore errors that
set returns.

I see no reason why we should throw when set fails, as we already do
hwc->perpare that will throw in case of a invalid state. In this case it
seem to be a normal behavior to ignore as we can't check if the overlay
is valid before committing. See:
https://android.googlesource.com/platform/frameworks/native/+/android-7.1.1_r28/services/surfaceflinger/SurfaceFlinger_hwc1.cpp#1314