DisplayGroup: allow commit() to fail for a few times
Commit() can fails for a transient reason. For example, if a client
disconnects between the content list is submitted and hwc->set() is
called, it will fails.

Usually, the next call should fix the problem, but if it happens more
than 3 times in a row, it may indicates that the problem is, in fact,
not transient. In this case, we'll forward the exception up the call
stack, which will then terminate the server.